Completion of fetchNewWord method. Now can get word from 3of6game.txt randomly. Unit tests are a little funky.

+++ HangmanQuest/HQGame.h
@@ -1,13 +1,19 @@
 #pragma once
 #include <string>
+#include <vector>
 #include <set>
+#include <cctype>
+#include <algorithm>
+#include <iostream>
+#include <fstream>
+#include <time.h>
 
 class HQGame
 {
 private:
 
 	std::string word; // The word to be guessed in Hangman.
-	std::set<char> letters; // A set of all the letters in the word.
+	std::set<char> letters; // A set of all the letters in the word to be guessed.
 	std::set<char> misses; // A set of guessed letters that were incorrect.
 	std::set<char> hits; // A set of guessed letters that were correct.
 
@@ -18,7 +24,7 @@ public:
 	void fetchNewWord(); // Fetches new word from 3of6game database.
 	int guess(char letter); // Guess that a letter is in a word.
 
-	int getWordLength() const; // Return number of letters in the word.
+	size_t getWordLength() const; // Return number of letters in the word.
 	std::string getWord() const; // Return word.
 	void setWord(std::string new_word); // Initializes gamestate with a word from the user.
 };
